fix `rgb2gray` function (return array in np.uint8)

+++ part-3-image-processing/image_utils.py
@@ -19,4 +19,4 @@ def rgb2gray(rgb: npt.NDArray[np.uint8]) -> npt.NDArray[np.uint8]:
         image in grayscale
 
     """
-    return np.dot(rgb[..., :3], [0.299, 0.587, 0.144])
+    return np.dot(rgb[..., :3], [0.299, 0.587, 0.144]).astype(np.uint8)
